[refactor]: 마이페이지 response DTO 구조 수정
[목적]: 각 정보를 객체로 구분하여 반환하기 위함.

[목표]: 마이페이지 응답 DTO에서 , , , 를 개별 객체로 구분하여 반환.

[달성도]:
  - 각 정보를 객체로 구분하여  구조 수정 완료.
  - 요구 사항을 반영하여 데이터 처리 방식 개선.
