+++ sequence_member/src/main/java/sequence/sequence_member/mypage/controller/MyPageController.java
@@ -3,22 +3,18 @@ package sequence.sequence_member.mypage.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.multipart.MultipartFile;
 import sequence.sequence_member.global.response.ApiResponseData;
 import sequence.sequence_member.global.response.Code;
-import sequence.sequence_member.member.dto.CustomUserDetails;
-import sequence.sequence_member.mypage.dto.MyActivityResponseDTO;
-import sequence.sequence_member.mypage.dto.MyPageRequestDTO;
-import sequence.sequence_member.mypage.dto.MyPageResponseDTO;
+import sequence.sequence_member.mypage.dto.MyPageRequestDto;
+import sequence.sequence_member.mypage.dto.MyPageResponseDto;
 import sequence.sequence_member.mypage.service.MyPageService;
 
 import java.util.List;
@@ -36,7 +32,7 @@ public class MyPageController {
         String username = SecurityContextHolder.getContext().getAuthentication().getName();
 
         try {
-            MyPageResponseDTO myPageDTO = myPageService.getMyProfile(username, page, size);
+            MyPageResponseDto myPageDTO = myPageService.getMyProfile(username, page, size);
             // 성공 응답 생성
             return ResponseEntity.ok(ApiResponseData.success(myPageDTO, "사용자 정보를 성공적으로 가져왔습니다."));
         } catch (Exception e) {
@@ -51,7 +47,7 @@ public class MyPageController {
 
     @PutMapping(value = "/api/mypage", consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.MULTIPART_FORM_DATA_VALUE})
     public ResponseEntity<ApiResponseData<String>> updateMyProfile(
-            @RequestPart(name = "myPageDTO") MyPageRequestDTO myPageDTO,
+            @RequestPart(name = "myPageDTO") MyPageRequestDto myPageDTO,
             @RequestPart(name = "authImgFile", required = false) MultipartFile authImgFile,
             @RequestPart(name = "portfolios", required = false) List<MultipartFile> portfolios
     ) {
@@ -74,7 +70,7 @@ public class MyPageController {
             @RequestParam(defaultValue = "10") int size  // 사이즈 기본값 10
     ) {
         try {
-            MyPageResponseDTO userProfile = myPageService.getUserProfile(nickname, page, size);
+            MyPageResponseDto userProfile = myPageService.getUserProfile(nickname, page, size);
             return ResponseEntity.ok(ApiResponseData.success(userProfile, nickname + "님의 정보를 성공적으로 가져왔습니다."));
         } catch (Exception e) {
             ApiResponseData errorResponse = ApiResponseData.failure(
