--- sequence_member/src/main/java/sequence/sequence_member/mypage/dto/MyArchiveDTO.java
@@ -1,13 +1,0 @@
-package sequence.sequence_member.mypage.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-@AllArgsConstructor
-public class MyArchiveDTO {
-    private List<PostDTO> writtenPosts;
-    private List<PostDTO> bookmarkedPosts;
-}
